Bug fixes and additions to Foundation classes
The add-on now auto adjusts every time a spell changes even by a
little, but is silent about it and only provides information to the
user when it’s necessary.
Fixed the message duplication bug by abstracting even further.
Fixed a bug with indefinitely long spells (such as aura of the
elements).
Updated localization files.
Corrected mage spells.